Add functionality to send and receive images in chat
