+++ src/agora/groupManager.js
@@ -1,12 +1,10 @@
 import {
     ChatGroupOptions,
     ChatMessage,
-    ChatManager,
     ChatMessageChatType,
     ChatMessageType,
     ChatConversationType,
-    ChatSearchDirection,
-    ChatGroupEventListener
+    ChatSearchDirection
 
 } from 'react-native-agora-chat';
 
@@ -19,54 +17,17 @@ export const getusername = async (chatClient) => {
     return user_name;
 };
 
-//Add Admin for a group
-export const makeadmin = async (isInitialized, chatClient, groupId, admin) => {
-    if (isInitialized === false || isInitialized === undefined) {
-        const message = 'Perform initialization first.';
-        console.log(message);
-        Alert.alert('Login Error', message);
+// Get Joined Groups
+export const getjoinedgroups = async (isInitialized, chatClient) => {
+    if ((isInitialized === false || isInitialized === undefined) && chatClient !== undefined) {
+        console.log('Perform initialization first.');
+        Alert.alert('Error', 'Perform initialization first.');
         return;
     }
-    return await chatClient.groupManager.addAdmin(groupId, admin)
-        .then(() => {
-            console.log("Admin Created Successfully")
-        })
-        .catch((error) => {
-            console.log("Error while creating admin")
-        });
+    let groups = await chatClient.groupManager.getJoinedGroups()
+    // console.log(groups)
+    return groups;
 }
-
-export const registerAdminAddedListener = (chatClient) => {
-    const groupListener = {
-        onAdminAdded: ({ admin, groupId }) => {
-            console.log(`Admin with ID: ${admin} has been added to group: ${groupId}`);
-        },
-        onMemberJoined: ({ member, groupId }) => {
-            console.log(`Member with ID: ${member} has joined group: ${groupId}`);
-        },
-    };
-    chatClient.groupManager.addGroupListener(groupListener)
-}
-
-export const registerMessageListener = (chatClient, groupId, setChats) => {
-    chatClient.chatManager.addMessageListener({
-        onMessagesReceived: (messages) => {
-            console.log('Received messages from admin:', messages);
-
-            messages.forEach(message => {
-                if (message.conversationId === groupId) {
-                    setChats(prevChats => [...prevChats, {
-                        id: message.msgId,
-                        message: message.body.content,
-                        sender: message.from
-                    }]);
-                }
-            });
-        }
-    });
-};
-
-
 
 //create a new group 
 export const creategroup = async (isInitialized, chatClient, groupName, groupDescription) => {
@@ -123,216 +84,179 @@ export const creategroup = async (isInitialized, chatClient, groupName, groupDes
     return group;
 };
 
-
-//get group from server 
-export const getgroup = async (chatClient, groupId) => {
-    let group = await chatClient.groupManager.fetchGroupInfoFromServer(
-        groupId,
-        false
-    )
-    console.log(group)
-    return group;
-}
-
-//addMembers(groupId: string, members: string[], welcome?: string): Promise<void>
-//Add group Members 
-export const addgroupmembers = async (isInitialized, chatClient, groupId, member) => {
-    if (isInitialized === false || isInitialized === undefined) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    // Check and log the type of members
-    console.log("Members: ", member);
-    console.log("Type of members: ", typeof member);
-    console.log("Is members an array?: ", Array.isArray(member));
-
-    // Convert members to an array of strings if it's not an array
-    if (Array.isArray(member)) {
-        // If already an array, ensure all elements are strings
-        member = member.map(member => String(member));
-    } else if (typeof member === 'string') {
-        // If members is a single string, convert it to an array
-        member = [member];
-    } else if (typeof member === 'object') {
-        // If members is an object, convert its values to an array of strings
-        member = Object.values(member).map(value => String(value));
-    } else {
-        Alert.alert('Error', 'Members should be an array, string, or an object.');
-        return;
-    }
-
-    console.log("Converted Members: ", member);
-
-    await chatClient.groupManager.addMembers(
-        groupId,
-        member,
-        "Welcome to the group"
-    )
-        .then(() => {
-            console.log("Members successfully added to the group.");
-            Alert.alert('Success', 'Members successfully added to the group.');
-        })
-        .catch((error) => {
-            console.error("Failed to add members to the group:", error);
-            Alert.alert('Error', `Failed to add members to the group: ${error.message}`);
+//Listener for received messages
+export const registerMessageListener = (chatClient, groupId, setChats) => {
+    chatClient.chatManager.addMessageListener({
+      onMessagesReceived: (messages) => {
+        console.log('Received messages from admin:', messages);
+  
+        messages.forEach(message => {
+          if (message.conversationId === groupId) {
+            let newChatItem;
+            const uniqueKey = `${message.msgId}-${message.localTime}`;
+  
+            if (message.body.type === 'txt') {
+              // Handle text message
+              newChatItem = {
+                id: message.localMsgId,
+                message: message.body.content,
+                sender: message.from,
+                type: 'txt'
+              };
+            } else if (message.body.type === 'img') {
+              // Handle image message
+              newChatItem = {
+                id: message.localMsgId,
+                message: '',
+                image: message.body.remotePath,
+                displayName: message.body.displayName,
+                width: message.body.width,
+                height: message.body.height,
+                sender: message.from,
+                type: 'img'
+              };
+            }
+  
+            setChats(prevChats => [...prevChats, newChatItem]);
+          }
         });
-}
-
-
-// Get Joined Groups
-export const getjoinedgroups = async (isInitialized, chatClient) => {
-    if ((isInitialized === false || isInitialized === undefined) && chatClient !== undefined) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    let groups = await chatClient.groupManager.getJoinedGroups()
-    // console.log(groups)
-    return groups;
-}
+      }
+    });
+  };
 
 //Receive Message from server
-export const createMessage = async (targetId) => {
-    const messageType = ChatMessageType.TXT;
+export const createMessage = async ({
+    groupId,
+    messageType,
+    message = null,
+    filePath = '',
+    displayName = '',
+    width = 100,
+    height = 100,
+    thumbnailLocalPath = '',
+    duration = 0,
+}) => {
     const chatType = ChatMessageChatType.GroupChat;
     let msg;
-    if (messageType === ChatMessageType.TXT) {
-        const content = "This is text message";
-        msg = ChatMessage.createTextMessage(targetId, content, chatType);
-    } else if (messageType === ChatMessageType.IMAGE) {
-        const filePath = "/data/.../image.jpg";
-        const width = 100;
-        const height = 100;
-        const displayName = "test.jpg";
-        msg = ChatMessage.createImageMessage(targetId, filePath, chatType, {
-            displayName,
-            width,
-            height,
-        });
-    } else if (messageType === ChatMessageType.CMD) {
-        const action = "writing";
-        msg = ChatMessage.createCmdMessage(targetId, action, chatType);
-    } else if (messageType === ChatMessageType.CUSTOM) {
-        const event = "gift";
-        const ext = { key: "value" };
-        msg = ChatMessage.createCustomMessage(targetId, event, chatType, {
-            params: JSON.stringify(ext),
-        });
-    } else if (messageType === ChatMessageType.FILE) {
-        const filePath = "data/.../foo.zip";
-        const displayName = "study_data.zip";
-        msg = ChatMessage.createFileMessage(targetId, filePath, chatType, {
-            displayName,
-        });
-    } else if (messageType === ChatMessageType.LOCATION) {
-        const latitude = "114.78";
-        const longitude = "39,89";
-        const address = "darwin";
-        msg = ChatMessage.createLocationMessage(
-            targetId,
-            latitude,
-            longitude,
-            chatType,
-            { address }
-        );
-    } else if (messageType === ChatMessageType.VIDEO) {
-        const filePath = "data/.../foo.mp4";
-        const width = 100;
-        const height = 100;
-        const displayName = "bar.mp4";
-        const thumbnailLocalPath = "data/.../zoo.jpg";
-        const duration = 5;
-        msg = ChatMessage.createVideoMessage(targetId, filePath, chatType, {
-            displayName,
-            thumbnailLocalPath,
-            duration,
-            width,
-            height,
-        });
-    } else if (messageType === ChatMessageType.VOICE) {
-        const filePath = "data/.../foo.wav";
-        const displayName = "bar.mp4";
-        const duration = 5;
-        msg = ChatMessage.createVoiceMessage(targetId, filePath, chatType, {
-            displayName,
-            duration,
-        });
-    } else {
-        throw new Error("Not implemented.");
+
+    switch (messageType) {
+        case ChatMessageType.TXT:
+            msg = ChatMessage.createTextMessage(groupId, message, chatType);
+            break;
+
+        case ChatMessageType.IMAGE:
+            msg = ChatMessage.createImageMessage(groupId, filePath, chatType, {
+                displayName,
+                width,
+                height,
+            });
+            break;
+
+        case ChatMessageType.FILE:
+            msg = ChatMessage.createFileMessage(groupId, filePath, chatType, {
+                displayName,
+            });
+            break;
+
+        case ChatMessageType.VIDEO:
+            msg = ChatMessage.createVideoMessage(groupId, filePath, chatType, {
+                displayName,
+                thumbnailLocalPath,
+                duration,
+                width,
+                height,
+            });
+            break;
+
+        case ChatMessageType.VOICE:
+            msg = ChatMessage.createVoiceMessage(groupId, filePath, chatType, {
+                displayName,
+                duration,
+            });
+            break;
+
+        default:
+            throw new Error("Not implemented.");
     }
+    console.log("Create Message: ",msg)
+    return msg;
 };
 
-//Send Message to server
-export const sendmsg = async (isInitialized, chatClient, targetId, content) => {
+
+export const sendmsg = async ({
+    isInitialized,
+    chatClient,
+    groupId,
+    messageType = null,
+    message = '',
+    filePath = '',
+    displayName = '',
+    width = 100,
+    height = 100,
+}) => {
     if (isInitialized === false || isInitialized === undefined) {
-        console.log('Perform initialization first.');  // Changed rollLog to console.log for consistency
+        console.log('Perform initialization first.');
         return;
     }
 
-    const messageType = ChatMessageType.TXT;
     let msg;
-
-    if (messageType === ChatMessageType.TXT) {
-        msg = ChatMessage.createTextMessage(
-            targetId,
-            content,
-            ChatMessageChatType.GroupChat
-        );
-    } else if (messageType === ChatMessageType.IMAGE) {
-        const filePath = "/data/.../image.jpg";
-        const width = 100;
-        const height = 100;
-        const displayName = "test.jpg";
-        msg = ChatMessage.createImageMessage(targetId, filePath, ChatMessageChatType.GroupChat, {
-            displayName,
-            width,
-            height,
-        });
+    try {
+        if (messageType === ChatMessageType.TXT) {
+            msg = await createMessage({ groupId, messageType, message });
+        } else if (messageType === ChatMessageType.IMAGE) {
+            // console.log("HERE");
+            msg = await createMessage({ groupId, messageType, filePath, displayName, width, height });
+        }
+    } catch (error) {
+        console.error("Failed to create message:", error.message);
+        return;
     }
 
-    const callback = {
-        onProgress: (localMsgId, progress) => {
-            console.log(`Sending message with ID ${localMsgId}: ${progress}% complete.`);
-        },
-        onError: (localMsgId, error) => {
-            console.log(`Error sending message with ID ${localMsgId}:`, error);
-        },
-        onSuccess: (message) => {
-            console.log(`Message sent successfully:`, message);
-        }
-    };
+    // Wrap sendMessage in a promise to handle the result from onSuccess
+    return new Promise((resolve, reject) => {
+        const callback = {
+            onProgress: (localMsgId, progress) => {
+                console.log(`Sending message with ID ${localMsgId}: ${progress}% complete.`);
+            },
+            onError: (localMsgId, error) => {
+                console.log(`Error sending message with ID ${localMsgId}:`, error);
+                reject(error);
+            },
+            onSuccess: (message) => {
+                console.log(`Message sent successfully:`, message);
+                // Update the `msg` with additional information from `message`
+                if (messageType === ChatMessageType.TXT) {
+                    resolve({
+                        id: msg.localMsgId,
+                        message: msg.body.content,
+                        sender: msg.from,
+                        type: 'txt'
+                    });
+                } else if (messageType === ChatMessageType.IMAGE) {
+                    resolve({
+                        id: msg.localMsgId,
+                        message: '',
+                        image: message.body.remotePath,  // Use the value from onSuccess
+                        displayName: message.body.displayName,
+                        width: message.body.width,
+                        height: message.body.height,
+                        sender: msg.from,
+                        type: 'img'
+                    });
+                }
+            }
+        };
 
-    return await chatClient.chatManager
-        .sendMessage(msg, callback)
-        .then(() => {
-            console.log('Send message: ' + msg.localMsgId);
-            const result = {
-                id: msg.localMsgId,
-                message: msg.body.content,
-                sender: msg.from
-            };
-            console.log(result)
-            return result;
-        })
-        .catch(reason => {
-            console.log('Send fail: ' + JSON.stringify(reason));
-        });
+        chatClient.chatManager
+            .sendMessage(msg, callback)
+            .catch(reason => {
+                console.log('Send fail: ' + JSON.stringify(reason));
+                reject(reason);
+            });
+    });
 };
 
 
-/*
-//Get the content of a message from message ID
-//1333463981527928442
-export const getmessage =  async (chatClient) => {
-    msgId = '1333463981527928442'
-    msg = await chatClient.chatManager.getMessage(msgId)
-    // message = await msg.getBody.ChatTextMessageBody.content;
-    const content = msg.body.content;
-    console.log(content);
-    // console.log(message);
-}
-*/
 
 //Receive Messages for a group chat with group id
 export const receivemessages = async (isInitialized, chatClient, targetId) => {
@@ -358,11 +282,34 @@ export const receivemessages = async (isInitialized, chatClient, targetId) => {
         .fetchHistoryMessages(targetId, ChatConversationType.GroupChat, params)
         .then((response) => {
             callback.onSuccess("HERE IS: ", response);
-            const chatData = response.list.map(message => ({
-                id: message.msgId,
-                sender: message.from,
-                message: message.body.content
-            }));
+            
+            // Handle different types of messages: text and image
+            const chatData = response.list.map(message => {
+                if (message.body.type === 'txt') {
+                    // Handle text message
+                    return {
+                        id: message.msgId,
+                        sender: message.from,
+                        message: message.body.content,
+                        type: 'txt'
+                    };
+                } else if (message.body.type === 'img') {
+                    // Handle image message
+                    return {
+                        id: message.msgId,
+                        sender: message.from,
+                        image: message.body.remotePath,
+                        type: 'img'  
+                    };
+                } else {
+                    return {
+                        id: message.msgId,
+                        sender: message.from,
+                        message: message.body.content || 'Unknown content',
+                        type: 'unknown'
+                    };
+                }
+            });
 
             return chatData;
         })
@@ -371,163 +318,3 @@ export const receivemessages = async (isInitialized, chatClient, targetId) => {
             return [];
         });
 };
-
-//Get all the information about group
-export const getGroupInfo = async (isInitialized, chatClient, groupId) => {
-    if (!isInitialized) {
-        rollLog('Perform initialization first.');
-        return;
-    }
-    const callback = new (class {
-        onSuccess() {
-            console.log('Group Info Received Successfully');
-        }
-        onError(error) {
-            console.log('Error receiving Group Info : ' + JSON.stringify(error));
-        }
-    })();
-    return await chatClient.groupManager.fetchGroupInfoFromServer(groupId, true)
-        .then((response) => {
-            callback.onSuccess(response);
-            const groupInfo = {
-                adminList: response.adminList,
-                blockList: response.blockList,
-                groupId: response.groupId,
-                description: response.description,
-                groupName: response.groupName,
-                memberCount: response.memberCount,
-                memberList: response.memberList,
-                muteList: response.muteList,
-                owner: response.owner,
-            };
-            console.log(groupInfo)
-            return groupInfo;
-        })
-        .catch((error) => {
-            callback.onError(error);
-            return [];
-        });
-}
-
-//removeMembers(groupId: string, members: string[]): Promise<void>
-export const removemember = async (isInitialized, chatClient, groupId, members) => {
-    if (!isInitialized) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    try {
-        await chatClient.groupManager.removeMembers(groupId, members);
-        console.log("Members successfully removed from the group.");
-        Alert.alert('Success', "Members successfully removed from the group.");
-    } catch (error) {
-        console.error("Failed to remove members from the group:", error);
-        Alert.alert('Error', `Failed to remove members from the group: ${error.message}`);
-    }
-}
-
-//blockMembers(groupId: string, members: string[]): Promise<void>
-export const blockmembers = async (isInitialized, chatClient, groupId, members) => {
-    if (!isInitialized) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    try {
-
-        await chatClient.groupManager.blockMembers(groupId, members);
-        console.log("Members successfully blocked from the group.");
-        Alert.alert('Success', "Members successfully blocked from the group.");
-    } catch (error) {
-        console.error("Failed to block members from the group:", error);
-        Alert.alert('Error', `Failed to block members from the group: ${error.message}`);
-    }
-}
-
-//muteMembers(groupId: string, members: string[], duration?: number): Promise<void>
-export const mutemembers = async (isInitialized, chatClient, groupId, members) => {
-    if (!isInitialized) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    try {
-        await chatClient.groupManager.muteMembers(groupId, members);
-        console.log("Members successfully muted from the group.");
-        Alert.alert('Success', "Members successfully muted from the group.");
-    } catch (error) {
-        console.error("Failed to mute members from the group:", error);
-        Alert.alert('Error', `Failed to mute members from the group: ${error.message}`);
-    }
-}
-
-//unblockMembers(groupId: string, members: string[]): Promise<void>
-export const unblockmembers = async (isInitialized, chatClient, groupId, members) => {
-    if (!isInitialized) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    try {
-        await chatClient.groupManager.unblockMembers(groupId, members);
-        console.log("Members successfully unblocked.");
-        Alert.alert('Success', "Members successfully unblocked.");
-    } catch (error) {
-        console.error("Failed to unblock members:", error);
-        Alert.alert('Error', `Failed to unblock members: ${error.message}`);
-    }
-}
-
-//unMuteMembers(groupId: string, members: string[]): Promise<void>
-export const unmutemembers = async (isInitialized, chatClient, groupId, members) => {
-    if (!isInitialized) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    try {
-        await chatClient.groupManager.unMuteMembers(groupId, members);
-        console.log("Members successfully unmuted.");
-        Alert.alert('Success', "Members successfully unmuted.");
-    } catch (error) {
-        console.error("Failed to unmute members:", error);
-        Alert.alert('Error', `Failed to unmute members: ${error.message}`);
-    }
-}
-
-//removeAdmin(groupId: string, admin: string): Promise<void>
-export const removeadmin = async (isInitialized, chatClient, groupId, members) => {
-    if (!isInitialized) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    try {
-        await chatClient.groupManager.removeAdmin(groupId, members);
-        console.log("Admin removed successfully.");
-        Alert.alert('Success', "Admin removed successfully.");
-    } catch (error) {
-        console.error("Failed to remove admin:", error);
-        Alert.alert('Error', `Failed to remove admin: ${error.message}`);
-    }
-}
-
-//leaveGroup(groupId: string): Promise<void>
-export const leavegroup = async (isInitialized, chatClient, groupId) => {
-    if (!isInitialized) {
-        console.log('Perform initialization first.');
-        Alert.alert('Error', 'Perform initialization first.');
-        return;
-    }
-    try {
-        await chatClient.groupManager.leaveGroup(groupId);
-        console.log("Group left successfully.");
-        Alert.alert('Success', "Group left successfully.");
-    } catch (error) {
-        console.error("Failed to leave group:", error);
-        Alert.alert('Error', `Failed to leave group: ${error.message}`);
-    }
-}
-
-
-
