+++ src/components/UploadFiles.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { TouchableOpacity, Image, Alert } from 'react-native';
 import DocumentPicker from 'react-native-document-picker';
 
-const UploadFiles = () => {
+const UploadFiles = ({setImageData}) => {
 
   const selectDoc = async () => {
     try {
@@ -10,12 +10,20 @@ const UploadFiles = () => {
         type: [DocumentPicker.types.allFiles],
         allowMultiSelection: true
       });
-      console.log(doc)
+      
+      const pickedImage = {
+        fileName: doc[0].name, 
+        fileSize: doc[0].size, 
+        fileType: doc[0].type, 
+        fileUri: doc[0].uri
+      }
+      console.log("Picked Image: ",pickedImage)
+      setImageData(pickedImage);
     } catch (err) {
       if (DocumentPicker.isCancel(err))
-        console.log("User Cancelled the Upload",err)
+        console.log("User Cancelled the Upload", err)
       else
-        console.log(error)
+        console.log(err)
     }
   }
 
@@ -31,7 +39,7 @@ const UploadFiles = () => {
       </TouchableOpacity> */}
 
       <TouchableOpacity
-        className="ml-2 bg-green-500 p-2 rounded-full"
+        className="ml-2 bg-blue-500 p-2 rounded-full"
         onPress={selectDoc}
       >
         <Image source={require('../../assets/icons/attach-file.png')}
